pyproject.toml
[project]
name = "questionnaires"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "selenium>=4.28.1",
    "requests>=2.32.3",
    "google-api-python-client>=2.167.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "pymysql>=1.1.1",
    "pandas>=2.3.0",
    "loguru>=0.7.3",
    "ratelimit>=2.2.1",
    "backoff>=2.2.1",
    "cryptography>=45.0.4",
    "pydantic>=2.12.2",
    "email-validator>=2.2.0",
    "tqdm>=4.67.1",
    "nameparser>=1.1.3",
    "pypdf2>=3.0.1",
    "inquirer>=3.4.1",
    "openpyxl>=3.1.5",
    "fastapi>=0.121.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "rich>=14.2.0",
    "pymupdf>=1.26.7",
]

[dependency-groups]
dev = [
    "icecream>=2.1.8",
    "ruff>=0.12.8",
]

[tool.ruff.lint]
select = ["UP", "FURB"]

[tool.ruff.lint.pydocstyle]
convention = "google"
